Fixed all mypy errors

+++ if_else/fraction_sum.py
@@ -5,8 +5,8 @@
 # 1 + (1/2) + (1/3) + (1/4) + (1/5) + ... + (1/n) 
 # You may assume that the parameter n is non-negative.
 
-def fraction_sum (num: int) -> int:
-    sum = 0
+def fraction_sum (num: int) -> float:
+    sum = 0.0
     for i in range (1, num + 1):
         sum += 1/i
     return sum
